As User I can subscribe to new comments notifications
I added checkbox to article page with proposal to subscribe to
comments email notifications. A separate model was created for it
with appropriate validations.

Change-Id: I46beeb062c7159612f831e8d11b092ddbbeaf72s

Story: 138
Task: 167

As User I can subscribe to new comments notification

I added checkbox with proposal to subscribe to comments email
notifications to article page. A separate model was created for this
with appropriate validations.

Change-Id: I05973e7f48d01835aa53f8096ed4f0ab95b0a659
Story: 138
Ticket: 167